+++ R/series__series.R
@@ -235,18 +235,20 @@ Series_str = method_as_active_binding(\() series_make_sub_ns(self, expr_str_make
 Series_struct = method_as_active_binding(\() series_make_sub_ns(self, expr_struct_make_sub_ns))
 
 
-# TODO: change the arguments to be match to Python Polars before 0.16.0
+# TODO: change the arguments in 0.17.0
 #' Create new Series
 #'
 #' This function is a simple way to convert basic types of vectors provided by base R to
 #' [the Series class object][Series_class].
 #' For converting more types properly, use the generic function [as_polars_series()].
+#' @param ... Treated as `values`, `name`, and `dtype` in order.
+#' In future versions, the order of the arguments will be changed to
+#' `pl$Series(name, values, dtype, ..., nan_to_null)` and `...` will be ignored.
 #' @param values any vector
 #' @param name Name of the Series. If `NULL`, an empty string is used.
 #' @param dtype One of [polars data type][pl_dtypes] or `NULL`.
 #' If not `NULL`, that data type is used to [cast][Expr_cast] the Series created from the vector
 #' to a specific data type internally.
-#' @param ... Ignored.
 #' @param nan_to_null If `TRUE`, `NaN` values contained in the Series are replaced to `null`.
 #' Using the [`$fill_nan()`][Expr_fill_nan] method internally.
 #' @return [Series][Series_class]
@@ -254,30 +256,46 @@ Series_struct = method_as_active_binding(\() series_make_sub_ns(self, expr_struc
 #' @seealso
 #' - [as_polars_series()]
 #' @examples
-#' # Constructing a Series by specifying name and values positionally:
-#' s = pl$Series(1:3, "a")
+#' # Constructing a Series by specifying name and values positionally (deprecated):
+#' s = suppressWarnings(pl$Series(1:3, "a"))
 #' s
 #'
 #' # Notice that the dtype is automatically inferred as a polars Int32:
 #' s$dtype
 #'
 #' # Constructing a Series with a specific dtype:
-#' s2 = pl$Series(1:3, "a", dtype = pl$Float32)
+#' s2 = pl$Series(values = 1:3, name = "a", dtype = pl$Float32)
 #' s2
 pl_Series = function(
-    values,
+    ...,
+    values = NULL,
     name = NULL,
     dtype = NULL,
-    ...,
     nan_to_null = FALSE) {
   uw = function(x) unwrap(x, "in pl$Series():")
+
+  if (!missing(...)) {
+    warning(
+      "The argument position of `pl$Series()` will be changed as of 0.17.0:\n",
+      "Until 0.17.0, the first argument corresponds to the values and the second argument to the name of the Series.\n",
+      "As of 0.17.0, the first argument will correspond to the name and the second argument to the values.\n",
+      " - 0.15.x: pl$Series(x, name, dtype, ..., nan_to_null)\n",
+      " - 0.17.0: pl$Series(name, values, dtype, ..., nan_to_null)\n",
+      "Since detecting the positional arguments now, use these arguments as `values`, `name` and `dtype`.\n",
+      "Use named arguments `values`, `name`, and `dtype` in `pl$Series()` instead of positional arguments to silence this warning."
+    )
+    dots = list(...)
+    values = values %||% dots[[1]]
+    if (length(dots) >= 2) name = name %||% dots[[2]]
+    if (length(dots) >= 3) dtype = dtype %||% dots[[3]]
+  }
 
   if (!is.null(dtype) && !isTRUE(is_polars_dtype(dtype))) {
     Err_plain("The dtype argument is not a valid Polars data type and cannot be converted into one.") |>
       uw()
   }
 
-  out = .pr$Series$new(values, name) |>
+  out = .pr$Series$new(name %||% "", values) |>
     uw()
 
   if (!is.null(dtype)) {