+++ NEWS.md
@@ -14,7 +14,26 @@
 ### Other breaking changes
 
 - Several functions have been rewritten to match the behavior of Python Polars.
-  - In `pl$Series()`, the first argument `x` is renamed to `values` (#933).
+  - In `pl$Series()` arguments are changed.
+    - The argument converted to the values of the Series is renamed from `x` to `values` (#933).
+    - Using positional arguments in `pl$Series()` is warned, since the
+      argument positions will be changed in the future (#966).
+      ```r
+      # polars 0.15.1 or earlier
+      # The first argument is `x`, the second argument is `name`.
+      pl$Series(1:3, "foo")
+      ```
+      ```r
+      # polars 0.17.0 or later (future version)
+      # The first argument is `values`, the second argument is `name`.
+      pl$Series("foo", 1:3)
+      ```
+      To silence the warning, please use named arguments.
+      ```r
+      # polars 0.16.0 or later
+      pl$Series(values = 1:3, name = "foo")
+      pl$Series(name = "foo", values = 1:3)
+      ```
   - `pl$implode(...)` is rewritten to be a syntactic sugar for `pl$col(...)$implode()` (#923).
   - Unify names of input/output function arguments (#935).
     - All arguments except the first argument must be named arguments.